Unify build_model* functions and move to asapdiscovery.ml.utils.
