+++ asapdiscovery/ml/scripts/sweep.py
@@ -18,313 +18,7 @@ from asapdiscovery.ml.scripts.train import (
     build_dataset,
     split_dataset,
 )
-from asapdiscovery.ml.utils import train
-
-
-def build_model(args, config=None):
-    """
-    Dispatch function for building the correct model and setting model_call
-    functions.
-
-    Parameters
-    ----------
-    args : dict
-        CLI arguments
-    config : dict, optional
-        Override wandb config
-
-    Returns
-    -------
-    Union[asapdiscovery.ml.models.GAT, mtenn.model.Model]
-        Build model
-    """
-
-    ## Correct model name if needed
-    model = args.model.lower()
-
-    ## Get config
-    if not config:
-        config = dict(wandb.config)
-
-    if model == "2d":
-        model = build_model_2d()
-        model_call = lambda model, d: torch.reshape(
-            model(d["g"], d["g"].ndata["h"]), (-1, 1)
-        )
-    elif (model == "schnet") or (model == "e3nn"):
-        import mtenn.conversion_utils
-        import mtenn.model
-
-        if model == "schnet":
-            model = build_model_schnet(config)
-            get_model = mtenn.conversion_utils.schnet.SchNet.get_model
-        else:
-            model_params = pkl.load(open(args.model_params, "rb"))
-            model = build_model_e3nn(100, *model_params[1:])
-            get_model = mtenn.conversion_utils.e3nn.E3NN.get_model
-
-        ## Take MTENN args from config if present, else from args
-        strategy = (
-            config["strat"].lower() if "strat" in config else args.strat.lower()
-        )
-        grouped = config["grouped"] if "grouped" in config else args.grouped
-
-        ## Check and parse combination
-        try:
-            combination = (
-                config["comb"].lower()
-                if "comb" in config
-                else args.comb.lower()
-            )
-            if combination == "mean":
-                combination = mtenn.model.MeanCombination()
-            elif combination == "boltzmann":
-                combination = mtenn.model.BoltzmannCombination()
-            else:
-                raise ValueError(f"Uknown value for -comb: {combination}")
-        except AttributeError:
-            ## This will be triggered if combination is left blank
-            ##  (None.lower() => AttributeError)
-            if grouped:
-                raise ValueError(
-                    f"A value must be provided for -comb if --grouped is set."
-                )
-            combination = None
-
-        ## Check and parse pred readout
-        try:
-            pred_readout = (
-                config["pred_r"].lower()
-                if "pred_r" in config
-                else args.pred_r.lower()
-            )
-            if pred_readout == "pic50":
-                pred_readout = mtenn.model.PIC50Readout()
-            elif pred_readout == "none":
-                pred_readout = None
-            else:
-                raise ValueError(f"Uknown value for -pred_r: {pred_readout}")
-        except AttributeError:
-            pred_readout = None
-
-        ## Check and parse comb readout
-        try:
-            comb_readout = (
-                config["comb_r"].lower()
-                if "comb_r" in config
-                else args.comb_r.lower()
-            )
-            if comb_readout == "pic50":
-                comb_readout = mtenn.model.PIC50Readout()
-            elif comb_readout == "none":
-                comb_readout = None
-            else:
-                raise ValueError(f"Uknown value for -comb_r: {comb_readout}")
-        except AttributeError:
-            comb_readout = None
-
-        ## Use previously built model to construct mtenn.model.Model
-        model = get_model(
-            model=model,
-            grouped=grouped,
-            strategy=strategy,
-            combination=combination,
-            pred_readout=pred_readout,
-            comb_readout=comb_readout,
-            fix_device=True,
-        )
-        model_call = lambda model, d: model(d)
-    else:
-        raise ValueError(f"Unknown model type: {model}")
-
-    return model, model_call
-
-
-def build_model_2d(config=None):
-    """
-    Build appropriate 2D graph model.
-
-    Parameters
-    ----------
-    config : Union[str, dict], optional
-        Either a dict or JSON file with model config options. If not passed,
-        `config` will be taken from `wandb`.
-
-    Returns
-    -------
-    asapdiscovery.ml.models.GAT
-        GAT graph model
-    """
-    from asapdiscovery.ml import GAT
-    from dgllife.utils import CanonicalAtomFeaturizer
-
-    if type(config) is str:
-        config = json.load(open(config_fn))
-    elif config is None:
-        config = wandb.config
-    elif type(config) != dict:
-        raise ValueError(f"Unknown type of config: {type(config)}")
-
-    # config.update({"in_node_feats": CanonicalAtomFeaturizer().feat_size()})
-    in_node_feats = CanonicalAtomFeaturizer().feat_size()
-
-    model = GAT(
-        in_feats=in_node_feats,
-        hidden_feats=[config["gnn_hidden_feats"]] * config["num_gnn_layers"],
-        num_heads=[config["num_heads"]] * config["num_gnn_layers"],
-        feat_drops=[config["dropout"]] * config["num_gnn_layers"],
-        attn_drops=[config["dropout"]] * config["num_gnn_layers"],
-        alphas=[config["alpha"]] * config["num_gnn_layers"],
-        residuals=[config["residual"]] * config["num_gnn_layers"],
-    )
-
-    return model
-
-
-def build_model_schnet(config=None):
-    """
-    Build appropriate SchNet model.
-
-    Parameters
-    ----------
-    config : Union[str, dict], optional
-        Either a dict or JSON file with model config options. If not passed,
-        `config` will be taken from `wandb`.
-
-    Returns
-    -------
-    mtenn.conversion_utils.SchNet
-        MTENN SchNet model created from input parameters
-    """
-    import mtenn.conversion_utils
-    from torch_geometric.nn import SchNet
-
-    ## Parse config
-    if type(config) is str:
-        config = json.load(open(config_fn))
-    elif config is None:
-        config = wandb.config
-    elif type(config) != dict:
-        raise ValueError(f"Unknown type of config: {type(config)}")
-
-    ## Get param values from config if they're there, otherwise just use default
-    ##  SchNet values
-    model_params = [
-        "hidden_channels",
-        "num_filters",
-        "num_interactions",
-        "num_gaussians",
-        "cutoff",
-        "max_num_neighbors",
-        "readout",
-    ]
-    model_params = {p: config[p] for p in model_params if p in config}
-
-    ## Build SchNet model and then MTENN model
-    model = SchNet(**model_params)
-    model = mtenn.conversion_utils.SchNet(model)
-
-    return model
-
-
-def build_model_e3nn(
-    n_atom_types,
-    num_neighbors,
-    num_nodes,
-    config=None,
-):
-    """
-    Build appropriate e3nn model.
-
-    Parameters
-    ----------
-    n_atom_types : int
-        Number off atom types in one-hot encodings. This will define the
-        dimensionality of the input into the model
-    num_neighbors : int
-        Approximate number of neighbor nodes that get convolved over for each
-        node. Used as a normalization factor in the model
-    num_nodes : int
-        Approximate number of nodes per graph. Used as a normalization factor in
-        the model
-    config : Union[str, dict], optional
-        Either a dict or JSON file with model config options. If not passed,
-        `config` will be taken from `wandb`.
-
-    Returns
-    -------
-    mtenn.conversion_utils.e3nn.E3NN
-        e3nn model created from input parameters
-    """
-    from e3nn.o3 import Irreps
-    import mtenn.conversion_utils
-
-    ## Parse config
-    if type(config) is str:
-        config = json.load(open(config_fn))
-    elif config is None:
-        config = wandb.config
-    elif type(config) != dict:
-        raise ValueError(f"Unknown type of config: {type(config)}")
-
-    ## Build hidden irreps
-    if "irreps_0o" in config:
-        irreps_hidden = Irreps(
-            [
-                (config["irreps_0o"], "0o"),
-                (config["irreps_0e"], "0e"),
-                (config["irreps_1o"], "1o"),
-                (config["irreps_1e"], "1e"),
-                (config["irreps_2o"], "2o"),
-                (config["irreps_2e"], "2e"),
-                (config["irreps_3o"], "3o"),
-                (config["irreps_3e"], "3e"),
-                (config["irreps_4o"], "4o"),
-                (config["irreps_4e"], "4e"),
-            ]
-        )
-    else:
-        irreps_hidden = Irreps(
-            [
-                (config["irreps_0"], "0o"),
-                (config["irreps_0"], "0e"),
-                (config["irreps_1"], "1o"),
-                (config["irreps_1"], "1e"),
-                (config["irreps_2"], "2o"),
-                (config["irreps_2"], "2e"),
-                (config["irreps_3"], "3o"),
-                (config["irreps_3"], "3e"),
-                (config["irreps_4"], "4o"),
-                (config["irreps_4"], "4e"),
-            ]
-        )
-
-    # input is one-hot encoding of atom type => n_atom_types scalars
-    # output is scalar valued binding energy/pIC50 value
-    # hidden layers taken from e3nn tutorial (should be tuned eventually)
-    # same with edge attribute irreps (and all hyperparameters)
-    # need to calculate num_neighbors and num_nodes
-    # reduce_output because we just want the one binding energy prediction
-    #  across the whole graph
-    model_kwargs = {
-        "irreps_in": f"{n_atom_types}x0e",
-        "irreps_hidden": irreps_hidden,
-        "irreps_out": "1x0e",
-        "irreps_node_attr": "1x0e" if config["lig"] else None,
-        "irreps_edge_attr": Irreps.spherical_harmonics(
-            config["irreps_edge_attr"]
-        ),
-        "layers": config["layers"],
-        "max_radius": config["max_radius"],
-        "number_of_basis": config["number_of_basis"],
-        "radial_layers": config["radial_layers"],
-        "radial_neurons": config["radial_neurons"],
-        "num_neighbors": num_neighbors,
-        "num_nodes": num_nodes,
-        "reduce_output": True,
-    }
-
-    return mtenn.conversion_utils.E3NN(model_kwargs=model_kwargs)
+from asapdiscovery.ml.utils import build_model, train
 
 
 def build_optimizer(model):
@@ -539,7 +233,15 @@ def main():
             ds_test = add_lig_labels(ds_test)
 
     ## Build model and set model call function
-    model, model_call = build_model(args)
+    model, model_call = build_model(
+        model_type=args.model,
+        e3nn_params=args.model_params,
+        strat=args.strat,
+        grouped=args.grouped,
+        comb=args.comb,
+        pred_r=args.pred_r,
+        comb_r=args.comb_r,
+    )
 
     ## Set up optimizer based on WandB config
     optimizer = build_optimizer(model)
